refactor: componentiza filtro de nome
